+++ coredis/client/basic.py
@@ -983,6 +983,10 @@ class Redis(Client[AnyStr]):
                     encoding=self._encodingcontext.get(),
                     disconnect_on_cancellation=should_block,
                 )
+                # TODO: Fix this! using both the release & should_block
+                #  flags to decide release logic is fragile. We should be
+                #  releasing early even in the cached response flow.
+
                 # if not blocking, no need to wait for reply
                 if not should_block:
                     released = True
@@ -1171,8 +1175,7 @@ class Redis(Client[AnyStr]):
         while True:
             with catch({WatchError: lambda _: logger.warning(msg)}):
                 async with self.pipeline(transaction=False) as pipe:
-                    if watches:
-                        await pipe.watch(*watches)
-                    return await func(pipe)
+                    async with pipe.watch(*watches):
+                        return await func(pipe)
             if watch_delay:
                 await sleep(watch_delay)